+++ app.js
@@ -1,18 +1,11 @@
 /* eslint-disable no-console */
 import express from 'express';
-import bodyParser from 'body-parser';
-import partyRouter from './src/routes/party-route';
-
+import index from './src/routes/index';
 // Initialize express
 const app = express();
 
-
-// Configure the middleware to accept both JSON and URL encoded data
-app.use(bodyParser.json());
-app.use(bodyParser.urlencoded({ extended: false }));
-
-// Module used to route the files relating to parties;
-app.use('/api/v1/parties', partyRouter);
+// Module used handle routes;
+app.use(index);
 
 
 const PORT = 5000;
